SCRUM-61 DRINK TABLES and fix the plate stucture

+++ repository/app/service/plate_service.py
@@ -32,10 +32,12 @@ def get_user_plates(id_user):
 
 def delete_Plate_service(userPlate_id):
     try:
-        # Referencia al documento del Plateo
         plate_ref = db.collection('Plate').document(userPlate_id)
+        # Check if the document exists before deleting
+        if not plate_ref.get().exists:
+            return {"error": "Plate not found"}
         plate_ref.delete()
-        return {"message": "user Plate  delete successful"}
+        return {"message": "User Plate deleted successfully"}
     except Exception as e:
         return {"error": str(e)}
